Reorganize all Concerns [fixes #163]
